requirements.txt
pyyaml
autodistill-grounding-dino
autodistill
opencv-python-headless
albumentations
pillow
rembg
numpy
requests
duckduckgo-search


